refactor part 2
