+++ lib/rx/configuration.rb
@@ -29,5 +29,22 @@ module Rx
         request_matcher: matcher
       )
     end
+
+    def connection
+      Faraday.new(base_path, headers: base_request_headers, request: request_options) do |conn|
+        conn.use :breakers
+        conn.request :json
+        # Uncomment this out for generating curl output to send to MHV dev and test only
+        # conn.request :curl, ::Logger.new(STDOUT), :warn
+
+        # conn.response :logger, ::Logger.new(STDOUT), bodies: true
+        conn.response :rx_parser
+        conn.response :snakecase
+        conn.response :raise_error
+        conn.response :json_parser
+
+        conn.adapter Faraday.default_adapter
+      end
+    end
   end
 end
