+++ lib/sm/client.rb
@@ -1,8 +1,6 @@
 # frozen_string_literal: true
-require 'faraday'
-require 'multi_json'
-require 'common/client/errors'
-require 'common/client/concerns/client_methods'
+
+require 'common/client/base'
 require 'common/client/concerns/mhv_session_based_client'
 require 'common/client/middleware/request/camelcase'
 require 'common/client/middleware/request/multipart_request'
@@ -19,7 +17,7 @@ require 'sm/api/messages'
 require 'sm/api/message_drafts'
 
 module SM
-  class Client
+  class Client < Common::Client::Base
     include Common::Client::MHVSessionBasedClient
     include SM::API::Sessions
     include SM::API::TriageTeams
@@ -27,24 +25,7 @@ module SM
     include SM::API::Messages
     include SM::API::MessageDrafts
 
-    def connection
-      @connection ||= Faraday.new(config.base_path, headers: config.base_request_headers, request: config.request_options) do |conn|
-        conn.use :breakers
-        conn.request :camelcase
-        conn.request :multipart_request
-        conn.request :multipart
-        conn.request :json
-        # Uncomment this out for generating curl output to send to MHV dev and test only
-        # conn.request :curl, ::Logger.new(STDOUT), :warn
-
-        # conn.response :logger, ::Logger.new(STDOUT), bodies: true
-        conn.response :sm_parser
-        conn.response :snakecase
-        conn.response :raise_error
-        conn.response :json_parser
-
-        conn.adapter Faraday.default_adapter
-      end
-    end
+    configuration SM::Configuration
+    session_klass SM::ClientSession
   end
 end
